@@ -3,15 +3,22 @@ package com.example.springoauth2.mvclogin.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.support.ResourceBundleMessageSource;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Configuration
 public class AppConfig {
 	
 	@Bean
-	public ResourceBundleMessageSource messageSource() {
+	ResourceBundleMessageSource messageSource() {
 		ResourceBundleMessageSource resource = new ResourceBundleMessageSource();
 		resource.setBasename("message");
 		return resource;
 	}
+	
+	 @Bean
+	 PasswordEncoder passwordEncoder() {
+	      return PasswordEncoderFactories.createDelegatingPasswordEncoder();
+	 }
 
 }
